+++ lib/src/core/service.dart
@@ -1,10 +1,9 @@
 library angel_framework.http.service;
 
 import 'dart:async';
-
 import 'package:angel_http_exception/angel_http_exception.dart';
 import 'package:merge_map/merge_map.dart';
-
+import 'package:quiver_hashcode/hashcode.dart';
 import '../util.dart';
 import 'anonymous_service.dart';
 import 'hooked_service.dart' show HookedService;
@@ -35,6 +34,9 @@ class Providers {
 
   /// Represents a request parsed from GraphQL.
   static const Providers graphQL = Providers(viaGraphQL);
+
+  @override
+  int get hashCode => hashObjects([via]);
 
   @override
   bool operator ==(other) => other is Providers && other.via == via;