+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Upgrade `http2` dependency.
 * Upgrade `uuid` dependency.
 * Fixed a bug that prevented body parsing from ever completing with `http2`.
+* Add `Providers.hashCode`.
 
 # 2.0.0-alpha.17
 * Revert the migration to `lumberjack` for now. In the future, when it's more